UX-460 refactor folder structure and files
