+++ packages/Textarea/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@paprika/textarea",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "description": "Textarea component",
   "author": "@paprika",
   "license": "MIT",
@@ -13,7 +13,8 @@
   },
   "dependencies": {
     "@babel/runtime-corejs2": "^7.3.1",
-    "@paprika/tokens": "^0.0.1",
+    "@paprika/stylers": "^0.1.0",
+    "@paprika/tokens": "^0.1.0",
     "class-names": "^1.0.0",
     "prop-types": "^15.7.2"
   },
